third_party/WebKit/LayoutTests/media/video-scales-in-media-document.html

<!DOCTYPE html>
<title>Test that video(media) documents scale to fit undersized containers.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<iframe style="width: 200px; height: 200px;"></iframe>
<script>
async_test(function(t) {
    var iframe = document.querySelector("iframe");

    iframe.onload = t.step_func(function() {
        var video = iframe.contentDocument.querySelector("video");
        video.oncanplaythrough = t.step_func_done(function() {
            assert_equals(video.offsetWidth, 200);
            assert_equals(video.offsetHeight, 164);
        });

        video.load();
    });

    iframe.src = "content/counting.ogv";
});
</script>
